requirements.txt
Django~=2.2.12
django-timezone-field~=4.0
djangorestframework~=3.11.0
djangorestframework-jwt~=1.11.0
pip~=20.1
pkg-resources~=0.0.0
PyJWT~=1.7.1
pytz~=2020.1
setuptools~=46.1.3
sqlparse~=0.3.1
wheel~=0.34.2